+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - docs: Added license
 - feat: Add admin management endpoints
+- feat: Add config to setup default admin accounts
 
 ### Changed
 - chore: Update dependencies